+++ src/python/pants/backend/docker/utils.py
@@ -4,9 +4,14 @@
 from __future__ import annotations
 
 import difflib
+import hashlib
+import json
 import os.path
+from collections import abc
+from dataclasses import asdict, is_dataclass
 from fnmatch import fnmatch
-from typing import Callable, Iterable, Iterator, Sequence, TypeVar
+from functools import partial
+from typing import Any, Callable, Iterable, Iterator, Sequence, TypeVar
 
 from pants.help.maybe_color import MaybeColor
 from pants.util.ordered_set import FrozenOrderedSet
@@ -164,3 +169,30 @@ def format_rename_suggestion(src_path: str, dst_path: str, *, colors: bool) -> s
     rem = color.maybe_red(src_path)
     add = color.maybe_green(dst_path)
     return f"{rem} => {add}"
+
+
+class JsonEncoder(json.JSONEncoder):
+    """Allow us to serialize everything, with a fallback on `str()` in case of any esoteric
+    types."""
+
+    def default(self, o):
+        """Return a serializable object for o."""
+        if is_dataclass(o):
+            return asdict(o)
+        if isinstance(o, abc.Mapping):
+            return dict(o)
+        if isinstance(o, abc.Sequence):
+            return list(o)
+        try:
+            return super().default(o)
+        except TypeError:
+            return str(o)
+
+
+json_dumps = partial(
+    json.dumps, indent=None, separators=(",", ":"), sort_keys=True, cls=JsonEncoder
+)
+
+
+def get_hash(value: Any, *, name: str = "sha256") -> hashlib._Hash:
+    return hashlib.new(name, json_dumps(value).encode("utf-8"))
